adv-manager/requirements.txt
# GBSC-Pro Register Editor Dependencies

# GUI Framework
PyQt6>=6.4.0

# Serial Communication
pyserial>=3.5

# Configuration Management
PyYAML>=6.0

# Development Dependencies (optional)
# pytest>=7.0.0
# pytest-qt>=4.0.0
# black>=22.0.0
# flake8>=4.0.0
